# services/auction-dutch/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation",
        "title": "Swagger",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://aban.io/support",
            "email": "info@aban.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/apis/admin/auctions/{id}/abort": {
            "post": {
                "description": "House aborts a non-terminal pre-settlement auction (e.g. threshold unmet). Emits auction.completed {final_state: ABORTED}. Illegal transition -\u003e RESOURCE_INVALID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auction-dutch-admin"
                ],
                "summary": "Abort auction (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auction UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AuctionResp"
                        }
                    },
                    "400": {
                        "description": "Illegal transition",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Auction not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/apis/admin/auctions/{id}/complete": {
            "post": {
                "description": "House completes a SETTLING auction. Emits auction.completed {final_state: COMPLETED}. Illegal transition -\u003e RESOURCE_INVALID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auction-dutch-admin"
                ],
                "summary": "Complete auction (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auction UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AuctionResp"
                        }
                    },
                    "400": {
                        "description": "Illegal transition",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Auction not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/apis/admin/auctions/{id}/open": {
            "post": {
                "description": "House opens a SCHEDULED auction. Requires at least one fully-locked, eligible participant. Sets open_at (the price clock origin) and emits auction.opened. Illegal transition -\u003e RESOURCE_INVALID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auction-dutch-admin"
                ],
                "summary": "Open auction (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auction UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AuctionResp"
                        }
                    },
                    "400": {
                        "description": "Illegal transition / no eligible participant",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Auction not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/apis/auctions/{id}": {
            "get": {
                "description": "Public read: the Dutch auction's state plus the SERVER-authoritative current price at now and the next drop instant. Clients render from these; the buy decision is always re-validated server-side.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auction-dutch"
                ],
                "summary": "Auction state + live price",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auction UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AuctionResp"
                        }
                    },
                    "400": {
                        "description": "Bad UUID",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/apis/auctions/{id}/buy": {
            "post": {
                "description": "The hammer action. The server re-computes the current price (a stale client price is ignored), validates the auction is OPEN and the caller is a fully eligible participant (KYC ∧ tier ∈ {MEMBER,VIP} ∧ deposit LOCKED ∧ full_lock LOCKED), then atomically transitions OPEN -\u003e HAMMER. The first valid buy wins; later buys return RESOURCE_INVALID. Emits auction.hammer.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auction-dutch"
                ],
                "summary": "Buy (hammer)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auction UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AuctionResp"
                        }
                    },
                    "400": {
                        "description": "Not OPEN / ineligible / already hammered",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Missing caller",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Auction not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/apis/auctions/{id}/lock": {
            "post": {
                "description": "Participant requests the 100% full lock before open. Creates a REQUESTED reservation and emits escrow.lock_requested {kind: FULL_LOCK}. Both a LOCKED deposit and a LOCKED full lock are required to buy.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auction-dutch"
                ],
                "summary": "Full lock (100%)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auction UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.ReservationResp"
                        }
                    },
                    "400": {
                        "description": "Bad UUID / ineligible / wrong state",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Missing caller",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Auction not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/apis/auctions/{id}/reserve": {
            "post": {
                "description": "Participant requests the 10% reservation deposit. Computes 10% of the ceiling, creates a REQUESTED reservation, and emits escrow.lock_requested {kind: DEPOSIT_10}. Caller identity + eligibility come from the gateway headers X-Account-Id / X-Account-Tier / X-Kyc-Approved.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auction-dutch"
                ],
                "summary": "Reserve (10% deposit)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auction UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.ReservationResp"
                        }
                    },
                    "400": {
                        "description": "Bad UUID / ineligible / wrong state",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Missing caller",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Auction not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AuctionResp": {
            "type": "object",
            "properties": {
                "ceilingCents": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "currentPriceCents": {
                    "description": "server-authoritative price(now)",
                    "type": "integer"
                },
                "dropIntervalSeconds": {
                    "type": "integer"
                },
                "dropStepCents": {
                    "type": "integer"
                },
                "floorCents": {
                    "type": "integer"
                },
                "hammerAt": {
                    "type": "string"
                },
                "hammerPriceCents": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "lotId": {
                    "type": "string"
                },
                "nextDropAt": {
                    "type": "string"
                },
                "openAt": {
                    "type": "string"
                },
                "state": {
                    "description": "DRAFT | ... | OPEN | HAMMER | ...",
                    "type": "string"
                },
                "winnerAccountId": {
                    "type": "string"
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.ReservationResp": {
            "type": "object",
            "properties": {
                "accountId": {
                    "type": "string"
                },
                "amountCents": {
                    "type": "integer"
                },
                "auctionId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "escrowRef": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "kind": {
                    "description": "DEPOSIT_10 | FULL_LOCK",
                    "type": "string"
                },
                "state": {
                    "description": "REQUESTED | LOCKED | RELEASED",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}